refactored utils/api file
