+++ client/src/features/createPost/components/CreatePostModal.jsx
@@ -3,8 +3,8 @@ import React, { useRef, useState } from 'react';
 import { useMutation, useQueryClient } from 'react-query';
 import useForm from '../../../hooks/useForm';
 import useOnClickOutside from '../../../hooks/useOnClickOutside';
-import { createPost, uploadImageCloud } from '../../../utils/api';
 import capitalize from '../../../utils/helpers';
+import { createPost, uploadImageCloud } from '../api';
 import CreatePostSelect from './CreatePostSelect';
 import ImageDropzone from './ImageDropzone';
 import ImagePreview from './ImagePreview';
