+++ client/src/utils/api.js
@@ -1,11 +1,4 @@
-import {
-  FEED_URL,
-  IMG_CLOUD_URL,
-  IMG_URL,
-  POST_URL,
-  SERVER_URL,
-  USER_URL,
-} from './constants';
+import { SERVER_URL, USER_URL } from './constants';
 
 export const handleApiResponse = async (response) => {
   const data = await response.json();
@@ -36,48 +29,6 @@ export const registerWithEmailAndPassword = async (data) => {
 
 export const fetchUser = async () => {
   return fetch(`${process.env.REACT_APP_SERVER_URL}api/users/user`, {
-    credentials: 'include',
-  }).then(handleApiResponse);
-};
-
-export const uploadImage = async (data) => {
-  const formData = new FormData();
-  formData.append('file', data[0]);
-  formData.append('name', data[0].name);
-
-  return fetch(`${SERVER_URL}${IMG_URL}`, {
-    method: 'POST',
-    credentials: 'include',
-    body: formData,
-  }).then(handleApiResponse);
-};
-
-export const uploadImageCloud = async (data) => {
-  return fetch(`${SERVER_URL}${IMG_CLOUD_URL}`, {
-    method: 'POST',
-    credentials: 'include',
-    headers: { 'Content-Type': 'application/json' },
-    body: JSON.stringify({ data }),
-  }).then(handleApiResponse);
-};
-
-export const createPost = async (data) => {
-  return fetch(`${SERVER_URL}${POST_URL}`, {
-    method: 'POST',
-    credentials: 'include',
-    headers: { 'Content-Type': 'application/json' },
-    body: JSON.stringify(data),
-  }).then(handleApiResponse);
-};
-
-export const getUserPosts = async () => {
-  return fetch(`${process.env.REACT_APP_SERVER_URL}${POST_URL}`, {
-    credentials: 'include',
-  }).then(handleApiResponse);
-};
-
-export const getFeedPosts = async () => {
-  return fetch(`${process.env.REACT_APP_SERVER_URL}${POST_URL}${FEED_URL}`, {
     credentials: 'include',
   }).then(handleApiResponse);
 };
